requirements.txt
streamlit==1.35.0
tensorflow==2.15.0
pillow==10.3.0
numpy==1.24.4
pandas==2.2.2